// src/app/compliance/models/Personalcompetente.ts
export interface Personalcompetente {
    orden: number;
    empleadoId: number;
    numEmpleado: string;
    nombre: string;
    apPaterno: string;
    apMaterno: string;
    genero: string;
    generoId: number;
    posicion: string;
    departamento: string;
    puesto: string;
    lugarDeTrabajo: string;
    lugarDeTrabajoId: number;
    usuarioModifico: string;
    fechaHoraUltimaModificacion: string;
    status: string;
    ver: string;
    editar: string;
    pdf: string;
    eliminar: string;
    nuevoexamen: string;
    mensajeEliminar: string;
}